Add and use array in G + extra acc options
Using allocatable arrays in G seems to have triggered the problem.

If I "$acc enter datain copy(G)" the field once per call, it seems to
work.

Outside of that, nothing seems to work, nor am I clear if it is supposed
to work.

I think this is the point where I need to seek advice.